detect only the relevant file changes when bumping a package version

+++ build/requiresTest.sh
@@ -1,20 +1,24 @@
 #!/bin/bash
+SHATESTED="$1"
 
-URLRUNS="https://api.github.com/repos/stjude/proteinpaint/actions/runs?status=success&branch=master&event=push"
-UNITRUNS=$(curl -sL -H "Accept: application/vnd.github+json" -H "X-GitHub-Api-Version: 2022-11-28" "$URLRUNS" > ./runs.json)
-SHA_IN_COMMIT=$(git rev-list -n10 HEAD)
-SHA_TESTED_LIST=$(node -p "(require('./runs.json')).workflow_runs?.filter(r => r.display_title != 'append release notes to change log').map(r => r.head_sha).join(' ')")
-rm runs.json
-for SHA in $SHA_TESTED_LIST; do
-	if [[ "$SHA_IN_COMMIT" ==  *"$SHA"* ]]; then
-		SHATESTED=$SHA
-		break
-	fi
-done
+if [[ "$SHATESTED" == "" ]]; then
+	URLRUNS="https://api.github.com/repos/stjude/proteinpaint/actions/runs?status=success&branch=master&event=push"
+	UNITRUNS=$(curl -sL -H "Accept: application/vnd.github+json" -H "X-GitHub-Api-Version: 2022-11-28" "$URLRUNS" > ./runs.json)
+	SHA_IN_COMMIT=$(git rev-list -n10 HEAD)
+	SHA_TESTED_LIST=$(node -p "(require('./runs.json')).workflow_runs?.filter(r => r.display_title != 'append release notes to change log').map(r => r.head_sha).join(' ')")
+	rm runs.json
+	for SHA in $SHA_TESTED_LIST; do
+		if [[ "$SHA_IN_COMMIT" ==  *"$SHA"* ]]; then
+			SHATESTED=$SHA
+			break
+		fi
+	done
+fi
 
 if [[ "$SHATESTED" == "" ]]; then
 	SHATESTED=$(git rev-parse HEAD~7)
 fi
+
 CHANGEDFILES="$(git diff --name-only HEAD $SHATESTED)"
 
 IS_CODE_OR_CONFIG="f => !f.endsWith('.md') && !f.endsWith('.txt') && !f.endsWith('ignore') && f != 'LICENSE' && f != 'DESCRIPTION'"
